+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.0
+
+- feat: add `StateNotifier` for simple use cases
+- **BREAKING**: add `ref` parameter for `ensureRef` callback
+
 ## 0.5.1
 
 - fix: lint fixes